Move network initialisations to HttpSender
Keep them in a single place, for easier replacement later.
